Refactored type parsers for clarity and to allow special casing of individual type parsers
