+++ ds3-autogen-net/src/main/java/com/spectralogic/ds3autogen/net/generators/parsers/element/BaseNullableElement.java
@@ -13,7 +13,7 @@
  * ****************************************************************************
  */
 
-package com.spectralogic.ds3autogen.net.generators.elementparsers;
+package com.spectralogic.ds3autogen.net.generators.parsers.element;
 
 import static com.spectralogic.ds3autogen.utils.Helper.capFirst;
