+++ ds3-autogen-net/src/main/java/com/spectralogic/ds3autogen/net/generators/parsers/element/NullableAttributeElement.java
@@ -13,28 +13,30 @@
  * ****************************************************************************
  */
 
-package com.spectralogic.ds3autogen.net.generators.elementparsers;
+package com.spectralogic.ds3autogen.net.generators.parsers.element;
 
 /**
- * Represents an element within a .net model that contains
- * a list of elements that do not have an encapsulating tag
+ * Represents an element within a .net model parser whose data is within
+ * an attribute of the current xml tag
  */
-public class NullableListElement extends BaseNullableElement {
+public class NullableAttributeElement extends BaseNullableElement {
 
-    public NullableListElement(
+    private final String parseAttributeFunc;
+
+    public NullableAttributeElement(
             final String name,
             final String xmlTag,
-            final String parserName) {
+            final String parserName,
+            final String parseAttributeFunc) {
         super(name, xmlTag, parserName);
+        this.parseAttributeFunc = parseAttributeFunc;
     }
 
     /**
-     * Gets the .net code for parsing this list of elements
+     * Gets the .net code for parsing this element from an element's attribute
      */
     @Override
     public String printParseElement() {
-        return getName() + " = element" +
-                ".Elements(\"" + getXmlTag() + "\")" +
-                ".Select(" + getParserName() + ").ToList()";
+        return getName() + " = " + getParserName() + "(element." + parseAttributeFunc + "(\"" + getXmlTag() + "\"))";
     }
 }
