+++ ds3-autogen-net/src/main/java/com/spectralogic/ds3autogen/net/generators/parsers/response/BaseResponseParserGenerator.java
@@ -13,7 +13,7 @@
  * ****************************************************************************
  */
 
-package com.spectralogic.ds3autogen.net.generators.parsermodels;
+package com.spectralogic.ds3autogen.net.generators.parsers.response;
 
 import com.google.common.collect.ImmutableList;
 import com.spectralogic.ds3autogen.api.models.Ds3Request;
