+++ ds3-autogen-net/src/main/java/com/spectralogic/ds3autogen/net/generators/parsers/response/ResponseParserModelGenerator.java
@@ -13,19 +13,13 @@
  * ****************************************************************************
  */
 
-package com.spectralogic.ds3autogen.net.generators.parsermodels;
+package com.spectralogic.ds3autogen.net.generators.parsers.response;
 
-import org.junit.Test;
+import com.spectralogic.ds3autogen.api.models.Ds3Request;
+import com.spectralogic.ds3autogen.api.models.Ds3Type;
+import com.spectralogic.ds3autogen.net.model.parser.BaseParser;
 
-import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertThat;
-
-public class JobListPayloadParserGenerator_Test {
-
-    private static final JobListPayloadParserGenerator generator = new JobListPayloadParserGenerator();
-
-    @Test
-    public void toNameToMarshal_Test() {
-        assertThat(generator.toNameToMarshal(null, null), is("Jobs"));
-    }
+@FunctionalInterface
+public interface ResponseParserModelGenerator<T extends BaseParser> {
+    T generate(final Ds3Request ds3Request, final String responsePayloadType, final String nameToMarshal);
 }
