+++ ds3-autogen-net/src/main/java/com/spectralogic/ds3autogen/net/model/typeparser/BaseTypeParserSet.java
@@ -13,19 +13,30 @@
  * ****************************************************************************
  */
 
-package com.spectralogic.ds3autogen.net.generators.parsermodels;
+package com.spectralogic.ds3autogen.net.model.typeparser;
 
-/**
- * Generates the BaseParser model for requests that have a JobList response payload.
- */
-public class JobListPayloadParserGenerator extends BaseResponseParserGenerator {
-    
-    /**
-     * The name to marshal value for the JobList payload is Jobs, which is not
-     * specified within the contract
-     */
-    @Override
-    public String toNameToMarshal(final String nameToMarshal, final String typeName) {
-        return "Jobs";
+import com.google.common.collect.ImmutableList;
+
+public class BaseTypeParserSet {
+
+    /** List of non-enum types (described by elements) */
+    private final ImmutableList<TypeParser> typeParsers;
+
+    /** List of enum type names */
+    private final ImmutableList<String> enumParsers;
+
+    public BaseTypeParserSet(
+            final ImmutableList<TypeParser> typeParsers,
+            final ImmutableList<String> enumParsers) {
+        this.typeParsers = typeParsers;
+        this.enumParsers = enumParsers;
+    }
+
+    public ImmutableList<TypeParser> getTypeParsers() {
+        return typeParsers;
+    }
+
+    public ImmutableList<String> getEnumParsers() {
+        return enumParsers;
     }
 }
