+++ ds3-autogen-net/src/test/java/com/spectralogic/ds3autogen/net/utils/NetNullableElementUtils_Test.java
@@ -15,7 +15,7 @@
 
 package com.spectralogic.ds3autogen.net.utils;
 
-import com.spectralogic.ds3autogen.net.generators.elementparsers.*;
+import com.spectralogic.ds3autogen.net.generators.parsers.element.*;
 import org.junit.Test;
 
 import static com.spectralogic.ds3autogen.net.utils.NetNullableElementUtils.*;
